safari-600.7-branch/LayoutTests/editing/inserting/insert-composition-whitespace-expected.txt

PASS compositingText is 'AB'
PASS confirmedText is 'AB'
PASS compositingText is 'A B'
PASS confirmedText is 'A B'
PASS compositingText is 'A  B'
PASS confirmedText is 'A  B'
PASS compositingText is 'A   B'
PASS confirmedText is 'A   B'
PASS compositingText is 'A    B'
PASS confirmedText is 'A    B'
PASS compositingText is ' AB'
PASS confirmedText is ' AB'
PASS compositingText is '  AB'
PASS confirmedText is '  AB'
PASS compositingText is '   AB'
PASS confirmedText is '   AB'
PASS compositingText is '    AB'
PASS confirmedText is '    AB'
PASS compositingText is '     AB'
PASS confirmedText is '     AB'
PASS compositingText is '      AB'
PASS confirmedText is '      AB'
PASS compositingText is '       AB'
PASS confirmedText is '       AB'
PASS compositingText is 'AB  '
PASS confirmedText is 'AB  '
PASS compositingText is 'AB   '
PASS confirmedText is 'AB   '
PASS compositingText is 'AB    '
PASS confirmedText is 'AB    '
PASS compositingText is 'AB     '
PASS confirmedText is 'AB     '
PASS compositingText is 'AB      '
PASS confirmedText is 'AB      '
PASS compositingText is 'AB       '
PASS confirmedText is 'AB       '
PASS compositingText is '  A  B  '
PASS confirmedText is '  A  B  '
PASS compositingText is '  A  B  '
PASS confirmedText is '  A  B  '
PASS compositingText is ' '
PASS confirmedText is ' '
PASS compositingText is '  '
PASS confirmedText is '  '
PASS compositingText is '   '
PASS confirmedText is '   '
PASS compositingText is 'AB'
PASS confirmedText is 'AB'
PASS compositingText is 'A B'
PASS confirmedText is 'A B'
PASS compositingText is 'A  B'
PASS confirmedText is 'A  B'
PASS compositingText is 'A   B'
PASS confirmedText is 'A   B'
PASS compositingText is 'A    B'
PASS confirmedText is 'A    B'
PASS compositingText is 'AB  '
PASS confirmedText is 'AB  '
PASS compositingText is 'AB   '
PASS confirmedText is 'AB   '
PASS compositingText is 'AB    '
PASS confirmedText is 'AB    '
PASS compositingText is 'AB     '
PASS confirmedText is 'AB     '
PASS compositingText is 'AB      '
PASS confirmedText is 'AB      '
PASS compositingText is 'AB       '
PASS confirmedText is 'AB       '
PASS compositingText is '  AB'
PASS confirmedText is '  AB'
PASS compositingText is '   AB'
PASS confirmedText is '   AB'
PASS compositingText is '    AB'
PASS confirmedText is '    AB'
PASS compositingText is '     AB'
PASS confirmedText is '     AB'
PASS compositingText is '      AB'
PASS confirmedText is '      AB'
PASS compositingText is '       AB'
PASS confirmedText is '       AB'
PASS compositingText is '  A  B  '
PASS confirmedText is '  A  B  '
PASS compositingText is '		A		B		'
PASS confirmedText is '		A		B		'
PASS compositingText is ' '
PASS confirmedText is ' '
PASS compositingText is '  '
PASS confirmedText is '  '
PASS compositingText is '   '
PASS confirmedText is '   '
PASS successfullyParsed is true

TEST COMPLETE
